- added logo + icons - changed tab bar icons - modified launch screen - added theme colors
