+++ Project Doggystyle iOS/Assets.xcassets/Logo/PlaceholderLogo.imageset/Contents.json
